Basic Structure of UVM Test bench with driver component
Creating a new class 'driver'   extending base or super class
Constructor function to initialize the 'driver' object when it is created.
task definition which represents the main functionality of the driver class.
Instantiating an object of the driver class
Setting the verbosity level of report ID
Calling the run() task of driver class which prints out the message