Digital Circuit to find the second largest value
A circuit that processes a clocked sequence of unsigned input values and outputs the second largest value seen in the sequence so far.
Circuit continuously monitor the input data ->maintains the record of 2 largest distinct values encountered so far in the sequence.
if only one unique value is seen, output will be 0;
encountering the same value multiple times doesn't change the ranking of values.