Rising edge Detector
A digital circuit that produces a single-cycle pulse on the output signal (data_out) one clock cycle after detecting a rising edge on the input signal (data_in). Additionally, when the reset signal (resetn) is asserted (active low), the value of data_in should be treated as 0, meaning no rising edge should be detected and dout should be 0.